requirements.txt
streamlit
python-dotenv
pandas
langchain
langchain_community
langchain_openai
faiss_cpu
openai